M2/18/backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
openai
python-dotenv